+++ src/constants.ts
@@ -1,14 +1,19 @@
 export const errorColor = 0xf23f42;
-export const mcColor = 0x6bc349;
 export const rerollButtonIds = {
     weapon: "reroll-weapon",
     mainWeapon: "reroll-mainweapon",
     subWeapon: "reroll-subweapon",
     specialWeapon: "reroll-specialweapon",
+    weaponType: "reroll-weapontype",
+    rule: "reroll-rule",
+    stage: "reroll-stage",
 };
-export const weaponCategoryName = {
+export const randomCategoryName = {
     weapon: "ブキ",
     mainWeapon: "メインウェポン",
     subWeapon: "サブウェポン",
     specialWeapon: "スペシャルウェポン",
+    weaponType: "ブキタイプ",
+    rule: "ルール",
+    stage: "ステージ",
 };