+++ src/index.ts
@@ -16,8 +16,6 @@ envs.forEach((value, key) => {
     }
 });
 
-import "./mc_listener";
-
 export class RainClient extends SapphireClient {
     public constructor() {
         super({
@@ -45,6 +43,7 @@ export class RainClient extends SapphireClient {
 declare module "@sapphire/pieces" {
     interface Container {
         database: PrismaClient;
+        lastGroupingResults: string[][] | undefined;
     }
 }