+++ src/interaction-handlers/reroll_rule.ts
@@ -6,11 +6,7 @@ import {
 import type { ButtonInteraction } from "discord.js";
 import { rerollButtonIds } from "../constants";
 import { RandomCommand } from "../commands/random";
-import {
-    checkVoiceChannelJoining,
-    getExecutedMember,
-    getRandomCommandOptions,
-} from "../utils/utils";
+import { checkVoiceChannelJoining, getExecutedMember } from "../utils/utils";
 
 @ApplyOptions<InteractionHandler.Options>({
     interactionHandlerType: InteractionHandlerTypes.Button,
@@ -19,18 +15,14 @@ export class ButtonHandler extends InteractionHandler {
     public async run(interaction: ButtonInteraction) {
         const member = await getExecutedMember(interaction);
         checkVoiceChannelJoining(member);
-        const options = await getRandomCommandOptions(interaction);
-        const replyOptions = await RandomCommand.buildRandomSubWeaponResult({
-            interaction,
-            options,
+        const replyOptions = await RandomCommand.buildRandomRuleResult({
             timestamp: true,
         });
         await interaction.update(replyOptions);
     }
 
     public override parse(interaction: ButtonInteraction) {
-        if (interaction.customId !== rerollButtonIds.subWeapon)
-            return this.none();
+        if (interaction.customId !== rerollButtonIds.rule) return this.none();
         return this.some();
     }
 }