+++ src/interaction-handlers/reroll_weapon.ts
@@ -5,26 +5,22 @@ import {
 } from "@sapphire/framework";
 import type { ButtonInteraction } from "discord.js";
 import { rerollButtonIds } from "../constants";
-import { RandomCommandOptions, UserCommand } from "../commands/random";
-import { getRandomCommandOptions } from "../utils/utils";
-import { CommandOptionFetchFailedError } from "../errors";
+import { RandomCommand } from "../commands/random";
+import {
+    checkVoiceChannelJoining,
+    getExecutedMember,
+    getRandomCommandOptions,
+} from "../utils/utils";
 
 @ApplyOptions<InteractionHandler.Options>({
     interactionHandlerType: InteractionHandlerTypes.Button,
 })
 export class ButtonHandler extends InteractionHandler {
     public async run(interaction: ButtonInteraction) {
-        let options: RandomCommandOptions;
-        try {
-            options = await getRandomCommandOptions(interaction);
-        } catch (error) {
-            if (error instanceof CommandOptionFetchFailedError) {
-                return;
-            } else {
-                throw error;
-            }
-        }
-        const replyOptions = await UserCommand.buildRandomWeaponResult({
+        const member = await getExecutedMember(interaction);
+        checkVoiceChannelJoining(member);
+        const options = await getRandomCommandOptions(interaction);
+        const replyOptions = await RandomCommand.buildRandomWeaponResult({
             interaction,
             options,
             timestamp: true,