+++ src/listeners/chat_input_command_error.ts
@@ -1,0 +1,76 @@
+import {
+    Events,
+    Listener,
+    ChatInputCommandErrorPayload,
+} from "@sapphire/framework";
+import {
+    errorEmbed,
+    generateItemNotFoundErrorEmbed,
+} from "../utils/embed_builder";
+import {
+    GuildMemberNotFoundError,
+    MemberVoiceChannelNotFoundError,
+    ItemNotFoundError,
+} from "../errors";
+
+export class ChatInputCommandErrorListener extends Listener {
+    public constructor(context: Listener.Context, options: Listener.Options) {
+        super(context, {
+            ...options,
+            event: Events.ChatInputCommandError,
+        });
+    }
+
+    public async run(
+        error: unknown,
+        { interaction }: ChatInputCommandErrorPayload
+    ) {
+        const embed = await (() => {
+            const subWeaponId =
+                interaction.options.getNumber("sub") ?? undefined;
+            const specialWeaponId =
+                interaction.options.getNumber("special") ?? undefined;
+            const seasonId =
+                interaction.options.getNumber("season") ?? undefined;
+            const weaponTypeId =
+                interaction.options.getNumber("weapontype") ?? undefined;
+            const single = interaction.options.getBoolean("single") ?? false;
+
+            const randomCommandOptions = {
+                subWeaponId,
+                specialWeaponId,
+                seasonId,
+                weaponTypeId,
+                single,
+            };
+
+            if (error instanceof GuildMemberNotFoundError) {
+                return errorEmbed(
+                    "実行したユーザーの情報が取得できませんでした。"
+                );
+            } else if (error instanceof MemberVoiceChannelNotFoundError) {
+                return errorEmbed(
+                    "ボイスチャンネルの情報が取得できませんでした。実行したサーバーでボイスチャンネルに参加しているか確かめてください。"
+                );
+            } else if (error instanceof ItemNotFoundError) {
+                return generateItemNotFoundErrorEmbed(randomCommandOptions);
+            } else if (error instanceof Error) {
+                return errorEmbed(`throw: ${error.message}`);
+            } else {
+                return errorEmbed(`throw: ${error}`);
+            }
+        })();
+
+        if (interaction.deferred || interaction.replied) {
+            return interaction.editReply({
+                content: null,
+                embeds: [embed],
+            });
+        }
+
+        return interaction.reply({
+            ephemeral: true,
+            embeds: [embed],
+        });
+    }
+}