+++ src/utils/utils.ts
@@ -8,8 +8,8 @@ import {
     CommandOptionFetchFailedError,
     GuildMemberNotFoundError,
     MemberVoiceChannelNotFoundError,
+    MemberVoiceChannelNotJoining,
 } from "../errors";
-import { errorEmbed } from "./embed_builder";
 import { RandomCommandOptions } from "../commands/random";
 
 export async function getExecutedMember(
@@ -20,10 +20,6 @@ export async function getExecutedMember(
             ? interaction.member
             : interaction.guild?.members.cache.get(interaction.user.id);
     if (!member) {
-        const embed = errorEmbed(
-            "実行したユーザーの`GuildMember`情報の取得に失敗しました。"
-        );
-        await interaction.reply({ embeds: [embed], ephemeral: true });
         throw new GuildMemberNotFoundError();
     } else {
         return member;
@@ -36,10 +32,6 @@ export async function getVoiceChannel(
     const member = await getExecutedMember(interaction);
     const voiceChannel = member.voice.channel;
     if (!voiceChannel) {
-        const embed = errorEmbed(
-            "実行したユーザーが参加している`VoiceBasedChannel`情報の取得に失敗しました。"
-        );
-        await interaction.reply({ embeds: [embed], ephemeral: true });
         throw new MemberVoiceChannelNotFoundError();
     } else {
         return voiceChannel;
@@ -63,11 +55,16 @@ export async function getVoiceChannelMembers({
     }
 }
 
+export function checkVoiceChannelJoining(member: GuildMember) {
+    const channel = member?.voice.channel;
+    if (!channel) {
+        throw new MemberVoiceChannelNotJoining();
+    }
+}
+
 export async function getRandomCommandOptions(interaction: ButtonInteraction) {
     const optionsJson = interaction.message.embeds[0].footer?.text;
     if (optionsJson == undefined) {
-        const embed = errorEmbed("オプションが取得できませんでした。");
-        interaction.reply({ embeds: [embed] });
         throw new CommandOptionFetchFailedError();
     }
     return JSON.parse(optionsJson) as RandomCommandOptions;